Move the move package from pkg to model
